Removed IController.ts and UserController class
