--- user/src/controllers/IController.ts
@@ -1,10 +1,0 @@
-import express from 'express';
-
-interface IController {
-    path: string;
-    router: express.Router;
-
-    initializeRoutes(): void;
-}
-
-export { IController };
